Java 8 language optimizations

+++ services/s3/src/main/java/software/amazon/awssdk/services/s3/checksums/ChecksumValidatingInputStream.java
@@ -105,12 +105,12 @@ public class ChecksumValidatingInputStream extends InputStream implements Aborta
         int read = -1;
 
         if (lengthRead < strippedLength) {
-            long maxRead = Math.min((long) Integer.MAX_VALUE, strippedLength - lengthRead);
-            int maxIterRead = (int) Math.min(maxRead, (long) len);
+            long maxRead = Math.min(Integer.MAX_VALUE, strippedLength - lengthRead);
+            int maxIterRead = (int) Math.min(maxRead, len);
 
             read = inputStream.read(buf, off, maxIterRead);
 
-            int toUpdate = (int) Math.min(strippedLength - lengthRead, (long) read);
+            int toUpdate = (int) Math.min(strippedLength - lengthRead, read);
 
             if (toUpdate > 0) {
                 checkSum.update(buf, off, toUpdate);
